[Refactor] chatbot page 추가
