+++ frontend/src/router/index.js
@@ -318,9 +318,10 @@ const routes = [
                 component: () => import('@/view/chat/ChatRoomPage.vue')
             },
             {
-                path: 'chatbot',
+                path: '/chatbot/:userId',
                 name: 'ChatBot',
-                component: () => import('@/view/chatbot/ChatBotPage.vue')
+                component: () => import('@/view/chatbot/ChatBotPage.vue'),
+                meta: { requiresAuth: true },
             }
         ]
     },
